feat(force-layout): WASM binding + fix wasm-test serde_json dev-dep
force-layout: extracts the d3-force-equivalent simulation loop into
`_force-layout-core` (pure compute, no deps). The napi crate is now a
thin shim translating FnodeSpec/FedgeSpec/SimulationOptions to/from
core types. WASM exposes `simulate(nodes, edges, options)` via
serde-wasm-bindgen.

Also fixes test-rust + wasm-test CI failures: the wasm-bindgen tests
for bm25, fuse, minisearch, force-layout, deepmerge use
`serde_json::Value` / `serde_json::json!` to round-trip JSON, but the
crate wasn't in the wasm sub-crate's dev-dependencies. Adding
`serde_json = "1"` to each makes the host-side `cargo test` and the
CI wasm-test job green again.
