+++ crates/_force-layout-core/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "amigo-force-layout-core"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
