+++ crates/bm25/wasm/Cargo.toml
@@ -15,6 +15,7 @@ serde = { version = "1", features = ["derive"] }
 
 [dev-dependencies]
 wasm-bindgen-test = { workspace = true }
+serde_json = "1"
 
 [package.metadata.wasm-pack.profile.release]
 wasm-opt = false
