+++ crates/deepmerge/wasm/Cargo.toml
@@ -18,6 +18,7 @@ serde_json = "1"
 
 [dev-dependencies]
 wasm-bindgen-test = { workspace = true }
+serde_json = "1"
 
 [package.metadata.wasm-pack.profile.release]
 # wasm-opt is invoked from a pre-built binaryen binary downloaded by
