+++ crates/force-layout/Cargo.toml
@@ -9,6 +9,7 @@ crate-type = ["cdylib"]
 [dependencies]
 napi = { workspace = true }
 napi-derive = { workspace = true }
+amigo-force-layout-core = { path = "../_force-layout-core" }
 
 [build-dependencies]
 napi-build = "2"
